requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
python-telegram-bot==20.6
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
